\name{alocv-package}
\alias{alocv-package}
\alias{alocv}
\docType{package}
\title{
\packageTitle{alocv}
}
\description{
\packageDescription{alocv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{alocv}
\packageIndices{alocv}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{alocv}

Maintainer: \packageMaintainer{alocv}
}
\references{
S. Wang, W. Zhou, H. Lu, A. Maleki, and V. Mirrokni, \emph{Leave-One-Out Approximatet for Fast Parameter Tuning in High Dimensions}, (2018), arXiv:1807.02694.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link[glmnet:glmnet-package]{glmnet}}
}
\examples{
~~ simple examples of the most important functions ~~
}
