% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocvBeta-package.R
\docType{package}
\name{alocvBeta-package}
\alias{alocvBeta}
\alias{alocvBeta-package}
\title{alocvBeta: Approximate Leave-One-Out Risk Estimation}
\description{
Computationally effcient approximation of the
    leave-one-out cross validation risk for (non)-smooth losses and
    regularizers.
}
\details{
Support elastic net regressions and SVM.
}
\references{
\itemize{
\item K. Rahnama Rad, A. Maleki, \emph{A scalable estimate of the extra-sample prediction error via approximate leave-one-out}, (2018), \url{https://arxiv.org/abs/1801.10243}.
\item S. Wang, W. Zhou, H. Lu, A. Maleki, and V. Mirrokni, \emph{Approximate Leave-One-Out Approximatet for Fast Parameter Tuning in High Dimensions}, (2018), \url{https://arxiv.org/abs/1807.02694}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Francis-Hsu/alocv}
  \item Report bugs at \url{https://github.com/Francis-Hsu/alocv/issues}
}

}
\author{
\strong{Maintainer}: Peng Xu \email{px2132@columbia.edu}

Authors:
\itemize{
  \item Linyun He \email{lh2870@columbia.edu}
  \item Wanchao Qin \email{wq2137@columbia.edu}
  \item Yuze Zhou \email{yuze.zhou@columbia.edu}
}

}
\keyword{internal}
