% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.glmnetALO.R
\name{plot.glmnetALO}
\alias{plot.glmnetALO}
\title{Plot the ALO curve produced by \code{glmnetALO}}
\usage{
\method{plot}{glmnetALO}(obj, ...)
}
\arguments{
\item{obj}{Fitted \code{\link{glmnetALO}} object.}

\item{...}{Other graphical parameters to plot.}
}
\description{
Plots the ALO curve, and upper and lower standard deviation curves, as a function of the \code{lambda} values used.
}
