\name{alocv-package}
\alias{alocv-package}
\alias{alocv}
\docType{package}
\title{
\packageTitle{alocv}
}
\description{
  \packageDescription{alocv}
}
\details{
  The DESCRIPTION file:
  \packageDESCRIPTION{alocv}
  \packageIndices{alocv}
}
\author{
  \packageAuthor{alocv}
  Maintainer: \packageMaintainer{alocv}
}
\references{
  S. Wang, W. Zhou, H. Lu, A. Maleki, and V. Mirrokni, \emph{Approximate Leave-One-Out Approximatet for Fast Parameter Tuning in High Dimensions}, (2018), \url{https://arxiv.org/abs/1807.02694}.
}
\keyword{alocv}
\seealso{
  \code{\link[glmnet:glmnet-package]{glmnet}}
}
