% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.glmnetALO.R
\name{coef.glmnetALO}
\alias{coef.glmnetALO}
\title{Extract the coefficients from a \code{glmnetALO} object}
\usage{
\method{coef}{glmnetALO}(obj, s = c("lambda.1se", "lambda.min"))
}
\arguments{
\item{obj}{Fitted \code{\link{glmnetALO}} object.}

\item{s}{Value of the penalty parameter \code{lambda} at which coefficients are required. 
Default is the value \code{s = "lambda.1se"}. Alternatively \code{s = "lambda.min"} can be used. 
If \code{s} is numeric, it is taken as the value of \code{lambda} to be used.}
}
\value{
An matrix of coefficients, no coefficients will be returned if no matching \code{s} is found.
}
\description{
This function extracts coefficients from a \code{glmnetALO} object for the optimal value chosen for \code{lambda}.
}
